requirements.txt
click==6.7
cycler==0.10.0
Flask==0.12.2
functools32==3.2.3.post2
gunicorn==19.7.1
import-string==0.1.0
itsdangerous==0.24
Jinja2==2.9.6
manage==0.1.13
MarkupSafe==1.0
matplotlib==2.0.2
netCDF4==1.2.9
numpy==1.13.1
pandas==0.20.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
scipy==0.19.1
six==1.10.0
subprocess32==3.2.7
urllib2==1498656401.94
Werkzeug==0.12.2
WTForms==2.1
